+++ pandas/core/frame.py
@@ -90,7 +90,7 @@ import pandas.core.nanops as nanops
 import pandas.core.ops as ops
 import pandas.formats.format as fmt
 from pandas.formats.printing import pprint_thing
-import pandas.tools.plotting as gfx
+import pandas.plotting.plotting as gfx
 
 import pandas.lib as lib
 import pandas.algos as _algos
@@ -5721,7 +5721,7 @@ DataFrame.hist = gfx.hist_frame
 @Appender(_shared_docs['boxplot'] % _shared_doc_kwargs)
 def boxplot(self, column=None, by=None, ax=None, fontsize=None, rot=0,
             grid=True, figsize=None, layout=None, return_type=None, **kwds):
-    import pandas.tools.plotting as plots
+    import pandas.plotting as plots
     import matplotlib.pyplot as plt
     ax = plots.boxplot(self, column=column, by=by, ax=ax, fontsize=fontsize,
                        grid=grid, rot=rot, figsize=figsize, layout=layout,
