+++ pandas/plotting/__init__.py
@@ -1,0 +1,1 @@
+from pandas.plotting.api import *           # noqa
