+++ pandas/tests/plotting/test_series.py
@@ -14,7 +14,7 @@ from pandas.util.testing import slow
 import numpy as np
 from numpy.random import randn
 
-import pandas.tools.plotting as plotting
+import pandas.plotting as plotting
 from pandas.tests.plotting.common import (TestPlotBase, _check_plot_works,
                                           _skip_if_no_scipy_gaussian_kde,
                                           _ok_for_gaussian_kde)
@@ -623,7 +623,9 @@ class TestSeriesPlots(TestPlotBase):
     @slow
     def test_kind_both_ways(self):
         s = Series(range(3))
-        for kind in plotting._common_kinds + plotting._series_kinds:
+        kinds = (plotting.plotting._common_kinds +
+                 plotting.plotting._series_kinds)
+        for kind in kinds:
             if not _ok_for_gaussian_kde(kind):
                 continue
             s.plot(kind=kind)
@@ -632,7 +634,7 @@ class TestSeriesPlots(TestPlotBase):
     @slow
     def test_invalid_plot_data(self):
         s = Series(list('abcd'))
-        for kind in plotting._common_kinds:
+        for kind in plotting.plotting._common_kinds:
             if not _ok_for_gaussian_kde(kind):
                 continue
             with tm.assertRaises(TypeError):
@@ -641,14 +643,14 @@ class TestSeriesPlots(TestPlotBase):
     @slow
     def test_valid_object_plot(self):
         s = Series(lrange(10), dtype=object)
-        for kind in plotting._common_kinds:
+        for kind in plotting.plotting._common_kinds:
             if not _ok_for_gaussian_kde(kind):
                 continue
             _check_plot_works(s.plot, kind=kind)
 
     def test_partially_invalid_plot_data(self):
         s = Series(['a', 'b', 1.0, 2])
-        for kind in plotting._common_kinds:
+        for kind in plotting.plotting._common_kinds:
             if not _ok_for_gaussian_kde(kind):
                 continue
             with tm.assertRaises(TypeError):
@@ -719,54 +721,57 @@ class TestSeriesPlots(TestPlotBase):
     def test_series_grid_settings(self):
         # Make sure plot defaults to rcParams['axes.grid'] setting, GH 9792
         self._check_grid_settings(Series([1, 2, 3]),
-                                  plotting._series_kinds +
-                                  plotting._common_kinds)
+                                  plotting.plotting._series_kinds +
+                                  plotting.plotting._common_kinds)
 
     @slow
     def test_standard_colors(self):
+        from pandas.plotting.style import _get_standard_colors
+
         for c in ['r', 'red', 'green', '#FF0000']:
-            result = plotting._get_standard_colors(1, color=c)
+            result = _get_standard_colors(1, color=c)
             self.assertEqual(result, [c])
 
-            result = plotting._get_standard_colors(1, color=[c])
+            result = _get_standard_colors(1, color=[c])
             self.assertEqual(result, [c])
 
-            result = plotting._get_standard_colors(3, color=c)
+            result = _get_standard_colors(3, color=c)
             self.assertEqual(result, [c] * 3)
 
-            result = plotting._get_standard_colors(3, color=[c])
+            result = _get_standard_colors(3, color=[c])
             self.assertEqual(result, [c] * 3)
 
     @slow
     def test_standard_colors_all(self):
         import matplotlib.colors as colors
+        from pandas.plotting.style import _get_standard_colors
 
         # multiple colors like mediumaquamarine
         for c in colors.cnames:
-            result = plotting._get_standard_colors(num_colors=1, color=c)
+            result = _get_standard_colors(num_colors=1, color=c)
             self.assertEqual(result, [c])
 
-            result = plotting._get_standard_colors(num_colors=1, color=[c])
+            result = _get_standard_colors(num_colors=1, color=[c])
             self.assertEqual(result, [c])
 
-            result = plotting._get_standard_colors(num_colors=3, color=c)
+            result = _get_standard_colors(num_colors=3, color=c)
             self.assertEqual(result, [c] * 3)
 
-            result = plotting._get_standard_colors(num_colors=3, color=[c])
+            result = _get_standard_colors(num_colors=3, color=[c])
             self.assertEqual(result, [c] * 3)
 
         # single letter colors like k
         for c in colors.ColorConverter.colors:
-            result = plotting._get_standard_colors(num_colors=1, color=c)
+            result = _get_standard_colors(num_colors=1, color=c)
             self.assertEqual(result, [c])
 
-            result = plotting._get_standard_colors(num_colors=1, color=[c])
+            result = _get_standard_colors(num_colors=1, color=[c])
             self.assertEqual(result, [c])
 
-            result = plotting._get_standard_colors(num_colors=3, color=c)
+            result = _get_standard_colors(num_colors=3, color=c)
             self.assertEqual(result, [c] * 3)
 
-            result = plotting._get_standard_colors(num_colors=3, color=[c])
+            result = _get_standard_colors(num_colors=3, color=[c])
             self.assertEqual(result, [c] * 3)
 
     def test_series_plot_color_kwargs(self):
