+++ setup.py
@@ -656,7 +656,8 @@ setup(name=DISTNAME,
                 'pandas.io.tests.sas',
                 'pandas.stats.tests',
                 'pandas.msgpack',
-                'pandas.util.clipboard'
+                'pandas.util.clipboard',
+                'pandas.plotting'
                 ],
       package_data={'pandas.io': ['tests/data/legacy_hdf/*.h5',
                                   'tests/data/legacy_pickle/*/*.pickle',
@@ -679,6 +680,7 @@ setup(name=DISTNAME,
                                   'tests/data/*.html',
                                   'tests/data/html_encoding/*.html',
                                   'tests/json/data/*.json'],
+                    'pandas.plotting': ['tests/data/*.csv'],
                     'pandas.tools': ['tests/data/*.csv'],
                     'pandas.tests': ['data/*.csv'],
                     'pandas.tests.formats': ['data/*.csv'],
